requirements.txt
appdirs==1.4.3
certifi==2018.1.18
chardet==3.0.4
Django==1.11
djangorestframework==3.6.3
flake8==3.5.0
gunicorn==19.7.1
idna==2.6
mccabe==0.6.1
olefile==0.44
packaging==16.8
pew==1.1.2
Pillow
psycopg2==2.7.1
pycodestyle==2.3.1
pyflakes==1.6.0
pyparsing==2.2.0
pytz==2017.2
requests==2.18.4
six==1.10.0
urllib3==1.22
